base pset finished

+++ pset3/kernel.cc
@@ -3,7 +3,7 @@
 #include "k-vmiter.hh"
 #include "obj/k-firstprocess.h"
 #include <atomic>
- 
+
 // kernel.cc
 //
 //    This is the kernel.
@@ -223,7 +223,7 @@ void process_setup(pid_t pid, const char* program_name) {
     ptable[pid].pagetable = kalloc_pagetable();
     assert(ptable[pid].pagetable);
     // Map kernel mem to user pagetable
-    for (uint64_t addr = 0; addr < PROC_START_ADDR; addr += PAGESIZE) {
+    for (uintptr_t addr = 0; addr < PROC_START_ADDR; addr += PAGESIZE) {
         vmiter k_pte = vmiter(kernel_pagetable, addr);
         int r = vmiter(ptable[pid].pagetable, addr)
             .try_map(k_pte.pa(), k_pte.perm());
@@ -499,7 +499,7 @@ pid_t syscall_fork() {
     ptable[pid].regs = current->regs;
     ptable[pid].state = P_RUNNABLE;
 
-    // Copy kernel mem mappings into new pagetable
+    // Copy parent's mem mappings into new pagetable
     for (vmiter it = vmiter(current->pagetable, 0); !it.done(); it.next()) {
         vmiter pte = vmiter(ptable[pid].pagetable, it.va());
         if (it.va() < PROC_START_ADDR || (it.user() && !it.writable())) {
